+++ src/components/Token/index.js
@@ -3,14 +3,15 @@ import PropTypes from 'prop-types';
 import Icon from 'react-oui-icons';
 import DismissButton from './DismissButton';
 import classNames from 'classnames';
+import { FILL_COLOR_MAP_LIGHT } from '../../utils/accessibility';
 
 const LIGHT_BACKGROUND_STYLES = ['tertiary'];
 
 // Light backgrounds require a dark font and dismiss Icon color.
-const getStylingInfo = style =>
-  LIGHT_BACKGROUND_STYLES.includes(style)
-    ? { fontClass: 'oui-token--font-dark', fillColor: 'black' }
-    : { fontClass: 'oui-token--font-light', fillColor: 'white' };
+const getStylingInfo = (style, backgroundColor) =>
+  LIGHT_BACKGROUND_STYLES.includes(style) || backgroundColor
+    ? { fontClass: 'oui-token--font-dark', dismissFillColor: 'black' }
+    : { fontClass: 'oui-token--font-light', dismissFillColor: 'white' };
 
 /**
  * Token to be used to make token lists.
@@ -20,6 +21,7 @@ const getStylingInfo = style =>
 const Token = ({
   description,
   dragHandleProps,
+  backgroundColor,
   hasSnugWrap,
   hasWrap,
   isDismissible,
@@ -33,32 +35,49 @@ const Token = ({
   testSection,
   usesDragHandle,
 }) => {
-  const classes = classNames({
+  const tokenContainerClasses = classNames({
     'oui-token-wrap': hasWrap && !hasSnugWrap,
     'oui-token-wrap--snug': hasSnugWrap,
     'oui-token-wrap--well': showWell,
     'oui-token--full-width': isFullWidth,
   });
+  const tokenClasses = classNames({
+    'oui-token': true,
+    [`oui-token--${style}`]: style,
+    'flex-align--center': !description,
+  });
   const tokenToolsClasses = classNames({
     'oui-token-tool': isDraggable || order,
     'cursor--grab': isDraggable,
   });
-  const { fontClass, fillColor } = getStylingInfo(style);
+  const { fontClass, dismissFillColor } = getStylingInfo(style, backgroundColor);
 
+  // ensure valid backgroundColor name
+  // (in the case that propType errors are ignored)
+  const tokenFillColor =
+    backgroundColor && Object.keys(FILL_COLOR_MAP_LIGHT).includes(backgroundColor)
+      ? FILL_COLOR_MAP_LIGHT[backgroundColor]
+      : null;
 
   if (usesDragHandle) {
     return (
       /* eslint-disable react/jsx-boolean-value */
       <div
         data-oui-component={ true }
-        className={ classes }
+        className={ tokenContainerClasses }
         data-test-section={ testSection }>
-        <div className={ `oui-token oui-token--${style}` }>
+        <div
+          className={ tokenClasses }
+          style={{ backgroundColor: tokenFillColor }}>
           <div className="flex flex-align--center">
             <div className={ tokenToolsClasses } data-token-handle>
-              {order && <span className="oui-token__number push--right">{order}</span>}
+              {order && (
+                <span className="oui-token__number push--right">{order}</span>
+              )}
               {isDraggable && (
-                <div className="oui-icon oui-token__move push--right oui-token__move--drag-handle" { ...dragHandleProps }>
+                <div
+                  className="oui-icon oui-token__move push--right oui-token__move--drag-handle"
+                  { ...dragHandleProps }>
                   <Icon name="hamburger" fill="#ffffff" />
                 </div>
               )}
@@ -73,8 +92,9 @@ const Token = ({
           {isDismissible && onDismiss && (
             <DismissButton
               onClick={ onDismiss }
-              fill={ fillColor }
+              fill={ dismissFillColor }
               testSection={ testSection }
+              backgroundColor={ backgroundColor }
             />
           )}
         </div>
@@ -87,38 +107,34 @@ const Token = ({
     /* eslint-disable react/jsx-boolean-value */
     <div
       data-oui-component={ true }
-      className={ classes }
+      className={ tokenContainerClasses }
       data-test-section={ testSection }>
-      <div
-        className={ tokenToolsClasses }
-        data-token-handle>
-        { order &&
-          <span className="oui-token__number push--right">
-            { order }
-          </span>
-        }
-        { isDraggable &&
+      <div className={ tokenToolsClasses } data-token-handle>
+        {order && (
+          <span className="oui-token__number push--right">{order}</span>
+        )}
+        {isDraggable && (
           <div className="oui-icon oui-token__move">
             <Icon name="ellipsis" fill="#c7c7c7" />
           </div>
-        }
+        )}
       </div>
-      <div className={ `oui-token oui-token--${style}` }>
+      <div
+        className={ tokenClasses }
+        style={{ backgroundColor: tokenFillColor }}>
         <div className={ fontClass }>
-          { name }
-          { description &&
-            <div className="oui-token__description">
-              { description }
-            </div>
-          }
+          {name}
+          {description && (
+            <div className="oui-token__description">{description}</div>
+          )}
         </div>
-        { (isDismissible && onDismiss) &&
+        {isDismissible && onDismiss && (
           <DismissButton
             onClick={ onDismiss }
-            fill={ fillColor }
+            fill={ dismissFillColor }
             testSection={ testSection }
           />
-        }
+        )}
       </div>
     </div>
     /* eslint-enable */
@@ -126,6 +142,11 @@ const Token = ({
 };
 
 Token.propTypes = {
+  /**
+   *  Color to use for the background of the token
+   */
+  backgroundColor: PropTypes.oneOf(Object.keys(FILL_COLOR_MAP_LIGHT)),
+
   /** Description explaining the token */
   description: PropTypes.string,
 
@@ -154,7 +175,7 @@ Token.propTypes = {
   isFullWidth: PropTypes.bool,
 
   /** Name label on token */
-  name: PropTypes.string.isRequired,
+  name: PropTypes.oneOf([PropTypes.string, PropTypes.node]).isRequired,
 
   /**
    * Function to call that dismisses the token. Required if the token is